fix(xpath): replace f64 with rust_decimal for xs:decimal precision
Migrate XdmAtomicValue::Decimal and ast::Literal::Decimal from f64 to
rust_decimal::Decimal (128-bit exact arithmetic). This fixes IEEE 754
rounding errors where expressions like
xs:decimal('0.1') + xs:decimal('0.2') eq xs:decimal('0.3')
incorrectly evaluated to false.

+++ crates/xpath/src/parser/ast.rs
@@ -1,9 +1,10 @@
 use compact_str::CompactString;
+use rust_decimal::Decimal;
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum Literal {
     Integer(i64),
-    Decimal(f64),
+    Decimal(Decimal),
     Double(f64),
     String(CompactString),
     Boolean(bool),
